requirements.txt
asgiref==3.5.2
contourpy==1.2.0
cycler==0.12.1
dj-database-url==0.5.0
Django==4.0.5
django-heroku==0.3.1
fonttools==4.47.0
gunicorn==20.1.0
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
psycopg2==2.9.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
zipp==3.17.0
